interfaces/seccomp/template.go: allow copy_file_range
This was recently introduced as an optimization to Go 1.15, and so apps that
start compiling may start to try and use it.

Note that Go 1.15 does now currently fall back to using other methods if copy_file_range
returns EPERM so that apps that get denied usage of copy_file_range with an EPERM 
will fallback to potentially slower implementations.

See https://github.com/golang/go/issues/40893 and
https://go-review.googlesource.com/c/go/+/249257/ for more details on the Go
issue and the fallback implementation.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>