systemd/systemd.go: support journald JSON messages with arrays for values
As per the man page,

Journal entries permit non-unique fields within the same log entry. JSON does not allow non-unique fields within objects. Due to this, if a non-unique field is encountered a JSON array is used as field value, listing all field values as elements.

Why journald thought it would be a good idea to support non-unique fields, who
knows, but we should handle this properly when we get messages from journald
this way.

Fixes: https://bugs.launchpad.net/snapd/+bug/1893279

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>