sysconfig/cloudinit.go: add AllowCloudInit and use GadgetDir for cloud.conf
ConfigureRunSystem now takes an option to specifically allow cloud-init to
remain enabled, with the intention that it is allowed to run only on first boot
of run mode, and that snapd in the devicemgr ensure will restrict or disable
cloud-init thereafter.

Additionally, if there is a cloud.conf in the GadgetDir provided to
ConfigureRunSystem, this is installed as trusted cloud-init config to a specific
filename in /etc/cloud/cloud.cfg.d.

Add unit tests for these scenarios.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>