overlord: add link participant for linkage transitions
The export manager needs to perform bookkeeping / updates on the new
current symbolic link that it is introducing. Instead of adding a new
task that is conceptually very similar to link-snap, and risking
breaking the complex sequence and dependencies between tasks, add a way
to hook into link-snap itself.

The link participant is a callback which can be used to obtain
notifications about link-snap, unlink-snap as well as their undo
executions.

Instead of carrying details as to how linkage changed, link participant
offers the system state and a snap name, allowing interested parties to
register a participant that can recover as little or as much information
as is required.

Since link-snap is so critical to the execution logic, link snap
participant failures are non-binding. The notification is only
dispatched after link-snap has effectively succeeded, save for making
the task as complete. Similar approach is used for unlink-snap as well
as the undo paths.

In effect a participant can be used to follow linkage changes and update
ancillary information, but cannot impact the overall change.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>