sysconfig/cloudinit.go: check for disallow cloud-init first, misc comments
The function installUnifiedCloudInitCfg is too generically named for it's
current implementation, so just rename it to something more specific for
now.

Additionally, add a TODO about filtering for the MAAS case, etc. as well
as update the doc-comment on ConfigureRunSystem to be more clear about
also implicitly configuring cloud-init via the gadget.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>