boot: keep track of the original asset when observing updates
When an observer is called during an update, it is provided with path to files
containing the original and new data applied by the update. This allows us to
restore the modeenv to correct state.

Specifically, a scenario in which an unexpected reboot may have occurred
mid-update, followed by a successful boot. The code marking boot successful,
will have processed modeenv and left hashes of assets that were present on disk
during boot, which could have been updated. When the update is picked up again
and an error occurs, rollback/cancelation would observe an incorrect state in
the modeenv and on the disk.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>