interfaces: update cups-control and add cups for providing snaps
* interfaces/cups-control: implement cups-control provided by snap

References:
https://forum.snapcraft.io/t/interface-request-cups-control-on-cups-snap-and-including-d-bus/15233/9

* interfaces: add cups companion interface to cups-control

On systems where the slot is provided by an app snap, the cups interface is
the companion interface to the cups-control interface. The design of these
interfaces is based on the idea that the slot implementation (eg cupsd) is
expected to query snapd on if the cups-control interface is connected or not
and the print service will mediate admin functionality (ie, the rules in
these interfaces allow connecting to the print service, but do not implement
enforcement rules; it is up to the print service to provide enforcement).

For now we limit this to app-providing slot implementations since there
are currently no classic systems with a mediating cupsd. We can revisit
this when it becomes available.

* cups-control: add some code comments and modernize profile comments

* comment adjustments

* add spread test

* fine-tune comments. Thanks to Ian Johnson

* cups_control_test.go: remove unnecessary TearDownTest()

* remove useless TearDownTest and simplify apparmor test. Thanks to Ian

* tests/main/interfaces-cups: only run on Ubuntu for strict-only tests

* disconnect cups-control earlier. Thanks to pedronis

* add base decl test for implicitOnCore plus app providers interfaces

* Clarify comments. Thanks to pedronis

Co-authored-by: Samuele Pedroni <pedronis@lucediurna.net>