tests: remove "set -e" from function only shell libs
The two shell libraries "nested.sh" and "reset.sh" currently set
"set -e" or "set -e -x". None of the other shell helper libraries
we have are doing this. Plus if one sources the library it may
change the current "set -e" setting. This is especially problematic
if nested.sh is imported to interactively debug a nested failure.
The first failed command will exit the shell and the session.

Hence this commit removes the "set -e" from the libraries, this
option should be set from whatever imports the shell helper.