bootloader/lkenv: fix typo
While this is exported, it is not currently used anywhere in Go, it merely
exists to keep API compatibility with the C code, which has an equivalent
definition for this, but does not have a typo in it.

This would be worse if the typo were in the C header file as we couldn't change
that without breaking clients like the various lk gadgets out there.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>