many: allow ignoring running apps for specific request
This patch adds support for "snap {install,refresh,revert}
--ignore-running" across API boundary, so that ultimately the two checks
for a snap being in use are not performed, even if the feature is
enabled.

This is primarily useful for spread tests, which assume the old behavior
and want to observe the in certain edge cases. It's not hidden as it may
be used by end users as well.

I left out snap {try,amend} even though they should have this option
as well, as none of the test required it. I'm looking for early feedback
on naming and implementation.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>