many/apparmor: adjust rule for reading apparmor profile for new kernel
Newer kernels such as the one in groovy currently will report AppArmor status
through /proc/<pid>/attr/apparmor/current instead of /proc/<pid>/attr/current
because due to some new LSM development, the latter is used by SELinux, so the
former is now used for AppArmor and libapparmor will search on the new path
to find the current profile if the kernel exposes it.

Also adjust snapd's apparmor package to use the new path if it exists.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>