overlord/snapstate: rewrite explanation of hard refresh
Samuele pointed out that there are two uses of the word "inhibit" that
are quite confusing. The hard refresh flow can inhibit application
startup and proceed with the refresh or can discover that there are
running applications and inhibit or postpone the refresh process.

I re-wrote that paragraph to make this distinction clearer.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>