packaging/{ubuntu,debian}: add liblzo2-dev as a dependency for building snapd
This is necessary to build squashfuse/snapfuse with lzo compression support,
which is necessary for snaps that now use lzo compression.

Not updating the non-ubuntu and non-debian package dependencies for now, as it's
unclear that using snapd inside non-ubuntu lxd containers is expected to work,
and that's the only known use-case for snapfuse right now.

Also update the lxd-snapfuse spread test to check that we can use lzo compressed
snaps inside an ubuntu lxd container now too.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>