daemon: write a maintenance.json file while shutting down
This file will be read by snapd clients to be able to introspect whether snapd
is alive or not, and avoids some instances of the opaque "connection refused"
errors we see when a snapd client tries to connect to the snapd daemon which has
already shut itself down for maintenance, either for a short restart or for a
reboot.

Because the client will eventually need to read this file, add it to the default
apparmor template so that snapctl can detect when snapd has shut down.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>