many: implement snap routine console-conf-start for synchronizing auto-refreshes
On first-boot, we often don't want refreshes to happen until the user has had a
chance to set their machine up with console-conf to claim ownership, so it's
desirable to expose a way for console-conf to instruct snapd to wait for
auto-refreshes until after console-conf is done.

The simplistic implementation here just sets refresh.hold to 20 minutes in the
future, and returns any in-progress autorefresh changes for the client to wait
on so if snapd is already auto-refreshing the user gets a message about that at
least.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>