spread-shellcheck: process paths from arguments in parallel
Attempt to speed up the spread-shellcheck execution a bit, by processing the
paths passed in the arguments in parallel. In particular, the run-checks command
executes: `spread-shellcheck spread.yaml tests`. The old code would firs process
spread.yaml, and then proceed to walk tests and process all files in
parallel (limited to by the workers count). I

t also happens that spread.yaml is composed of many longish sections and the
processing of a complete file takes longer than for a typical test.

The patch makes all the locations passed in the command line be processed in
parallel by one executors pool.

A bit of data, running `spread-shellcheck spread.yaml tests` on my machine:
- baseline:       75s
- with the patch: 59s

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>