spread-shellcheck: increase timeout to 30sec
spread-shellcheck imposes a timeout on each shellcheck instance that
runs, so that it automatically stops and fails after 10 seconds. This
was done long ago, IIRC, to allow shellcheck to not run forever on
malformed input.

With the increased concurrency, on commonly used machines using
symmetric multi-threading, the OS sees virtual CPU cores that are
suitable for concurrency at a thread level, not at a process level.
The extra system load from the overall parallelism is now pushing the
execution of the slowest shellcheck, at least on a Intel Haswell-era
mobile core i7, past the 10 second limit.

Since the limit is pretty much entirely arbitrary, increase it to 30
seconds.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>