cmd: hide the --ignore-running option
The upcoming hook for blocking refreshes is not yet decided but once it
is in place, there may be a separate option to ignore that hook, or both
overrides may be fused into one. Until we look at how users will
interact, it's better to start with a hidden, unstable option name.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>