daemon/daemon.go: revert restarting/needsFullReboot change
We previously would only send STOPPING=1 to systemd (as well as some other small
behavior changes inside Stop), if the system was _not_ rebooting, but the
previous commit broke this.

Instead send STOPPING=1 whenever we are stopping and we are not doing a full
system reboot.

Also update a unit test to make sure this behavior is tested now.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>