tests/nested/core20/kernel-reseal: use the repacked test kernel
We cannot use the pc-kernel directly from the store as it may have a version of
snap-bootstrap that will be unable to unseal the keys and unlock the encrypted
volumes. Instead, use the version that kernel snap that was repacked with the
version of snap-boostrap from the current source tree.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>