gadget/many: drop usage of gpt attr 59 for indicating creation of partitions
Usage of GPT attribute 59 technically requires ownership of the GPT Type UUID we
are using, which is just "Linux filesystem data" and not arguably owned by any
one entity.

Additionally, it is difficult to set this bit on tools traditionally used by arm
devices for flashing the image on them where all partitions are needed to be
present (even if empty) when flashing the internal memory.

As such, drop usage of the attribute and instead just perform matching based on
the labels that we expect to create during installation, as well as the strict
validation we do for the on-disk structures matching what is in the gadget.yaml.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>