gadget/internal: let mkfs.ext4 figure out the best setup
The mkfs.ext4 use the heuristic for choosing the best setup. It should check the
block device size and choose and appropriate setting from /etc/mke2fs.conf (or
the built-in copy if the file is missing).

In the particular case of ubuntu-save, it will choose the 'small' profile, which
should give us more free space on the device.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>