secboot: refactor Unlock...IfEncrypted to take keyfile + check disks first
Refactor UnlockVolumeUsingSealedKeyIfEncrypted to always check the disk for
partitions first, then do the TPM connection. This is so that we always know
for sure (if we can know and FindMatchingPartitionUUID doesn't have a
catastrophic error) when we return from UnlockVolumeUsingSealedKeyIfEncrypted
that there is an encrypted device or not. This will make the degraded
recover implementation more robust as it means that even with TPM errors
for i.e. unlocking ubuntu-data-enc, we will still know that we should try
to unlock ubuntu-save.

Additionally, we will want to be able to call this function with different
sealed keys for ubuntu-data, either the main run object or the fallback
object, both of which will be in the encryption dir on ubuntu-seed. As such,
take in a specific key to use rather than construct the name of the keyfile.

Finally, update usage of UnlockVolumeUsingSealedKeyIfEncrypted in 
snap-bootstrap to account for this change.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>