cmd/s-b/initramfs-mounts: refactor recover mode to implement degraded mode
Refactor `snap-bootstrap initramfs-mounts`'s recover mode logic to use a state
machine which implements degraded mode for recover mode. This means that if
mounting data with the run key fails, we try to use the fallback key, and
similarly for ubuntu-save. See the state diagram in the code for full
details of all the transitions.

This also means that we now expect to never call
UnlockVolumeUsingSealedKeyIfEncrypted with lockKeysOnFinish set to true
from recover mode, and instead will always use LockTPMSealedKeys()
instead because it's not clear deterministically which call to Unlock...
will be the last one.

Additionally, add unit full-scale tests for all the relevant scenarios
in recover mode. We also eventually will want to test each state inside
the state machine on it's own to ensure that it is transitioning as
expected, etc.

Thanks to @bboozzoo for suggesting this pattern and helping implement
it.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>