cmd/s-b/initramfs-mounts: refactor state machine UnlockResult handling
Refactor similar to 5f69541e5c4096931b0f91b7d86b3825b434cddd, but for setting
the UnlockState and UnlockKey, etc. when trying to unlock with the run key or
fallback keys for ubuntu-data and ubuntu-save.

This makes the state machine's steps more readable as it is more clear what they
transition to, leaving the details of what needs to be updated in the state
machine to helpers. It also makes it more clear that unlocking ubuntu-save with
the fallback key is doing essentially the same things as unlocking ubuntu-data
with the fallback key.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>