secboot: call BlockPCRProtectionPolicies even if the TPM is disabled
BlockPCRProtectionPolicies inserts PCR events that signal the
transition from initramfs to runtime, in the same way that EV_SEPARATOR
events signal the transition from pre-OS (platform firmware) to
OS present. When the TPM is disabled, platform firmware will cap
PCRs 0-7 with a EV_SEPARATOR event. There is no reason for us not
to behave in a similar way.