snap-bootstrap,secboot: call BlockPCRProtectionPolicies in all boot modes
BlockPCRProtectionPolicies is only called during run and recover modes.
This changes snap-bootstrap so that it is called in all boot modes using
a deferred function that is declared very early which ensures that it is
called on most early failures as well.

In addition to this, the same code path for calling it is now used on
all boot modes, which allows UnlockVolumeUsingSealedKeyIfEncrypted to
be significantly simplified.