secboot,cmd/snap-bootstrap: refactor UnlockResult handling to distinguish cases
Now UnlockResult has two better named fields, FsDevice and PartDevice, where
the PartDevice is always the underlying block device, and FsDevice (if not
the empty string) is whatever device should be mounted to access the
filesystem.

Also update the handling of setting up state for degraded mode in recover
mode to use these variables in a simpler way.

Finally, simplify some logic in the state machine around ubuntu-save being
optional if it is unencrypted and add an additional consistency check that
if ubuntu-data was mounted unencrypted that ubuntu-save must be unencrypted
as well.

Thanks to @pedronis for suggesting this direction of the code and providing
example code.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>