gadget/gadget.go: allow system-recovery-{image,select} as roles in gadget.yaml
This is for littlekernel bootloaders, where we previously only had
system-boot-select, we now also need system-recovery-select for UC20 systems.

Similarly, we are introducing a system-recovery-image, which presently will
not be treated any differently from system-boot-image, but will be used to store
specifically kernels for recovery systems, so we may find it desirable to
distinguish between the two kinds of boot roles.

Also document the roles a bit more.

The specific gadget.yaml here is from the reference dragonboard 410c which was
"quasi-enabled" for UC20 with LK using the attached gadget.yaml.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>